// libs/firefly-shared/src/state/child/user-subscriptions/user-subscriptions.actions.enum.ts
export enum ActionsUserSubscriptions {
  Reset = '[UserSubscriptions] Reset',
  GetData = '[UserSubscriptions] Get Data',
  SetData = '[UserSubscriptions] Set Data',
  Get = '[UserSubscriptions] Get',
  Add = '[UserSubscriptions] Add',
  Remove = '[UserSubscriptions] Remove',
  Sync = '[UserSubscriptions] Sync',
  Filter = '[UserSubscriptions] Filter'
}